+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * GraphQL: Add page-based pagination (#3175, #3517)
 * GraphQL: Possibility to add a custom description for queries, mutations and subscriptions (#3477, #3514)
 * GraphQL: Support for field name conversion (serialized name) (#3455, #3516)
+* GraphQL: **BC** `operation` is now `operationName` to follow the standard (#3568)
 * OpenAPI: Add PHP default values to the documentation (#2386)
 * Deprecate using a validation groups generator service not implementing `ApiPlatform\Core\Bridge\Symfony\Validator\ValidationGroupsGeneratorInterface` (#3346)
 
@@ -34,7 +35,6 @@
 * HTTP: Location header is only set on POST with a 201 or between 300 and 400 #3497
 * GraphQL: Do not allow empty cursor values on `before` or `after` #3360
 * Bump versions of Swagger UI, GraphiQL and GraphQL Playground #3510
->>>>>>> 2.5
 
 ## 2.5.4